Home Page Update
Changed the displayed model on the home page. (New model file is too large to push)

Changed multiple camera and scale values in the 3d handler js file for better cohesion.

Moved the toggle button to the top of the page.

Limited the page height on the home page to 100% with no overflow.